refactoring

+++ Repeat/Pages/Administration/Questions/Details.cshtml.cs
@@ -26,7 +26,7 @@ namespace Repeat.Pages.Administration.Questions
                 return NotFound();
             }
 
-            this.Question = await _questionService.GetQuestionByIdAsync(id);
+            this.Question = await _questionService.GetQuestionByIdAsync((int) id);
 
             if (this.Question == null)
             {
